+++ game/models.py
@@ -128,6 +128,35 @@ class Board:
                     row_str += str(cell.letter.letter) + " "
             print(row_str)
 
+    def calculate_word_value(self, word:list):
+        
+        counter = 0
+        word_multiplier = 1
+
+        for i in range(len(word)):
+            cell = word[i]
+
+            if cell.letter is None:
+                    return 0
+            
+            if cell.state == False:
+                 counter = counter + cell.letter.value
+            
+            if cell.multiplier_type == 'letter' and cell.state == True:
+                counter = counter + (cell.letter.value * cell.multiplier * word_multiplier)
+
+            if cell.multiplier_type == 'word' and word_multiplier != 1 and cell.state == True:
+                counter = (counter + (cell.letter.value * word_multiplier))
+                word_multiplier = cell.multiplier
+                counter = counter * word_multiplier
+
+            if cell.multiplier_type == 'word' and word_multiplier == 1 and cell.state == True:
+                word_multiplier = cell.multiplier
+                counter = ((counter + cell.letter.value) * word_multiplier)
+            cell.used_cell(cell)
+        return counter    
+
+
 class Cell:
     def __init__(self, multiplier = 1, multiplier_type = 'letter', letter = None, state = True):
         self.multiplier = multiplier
@@ -151,4 +180,4 @@ class Cell:
         self.multiplier_type = 'letter'
         self.state = False
 
-
+