+++ game/value.py
@@ -1,29 +1,3 @@
 from game.models import Cell, Tiles
 
-def calculate_word_value(word):
-        
-        counter = 0
-        word_multiplier = 1
-
-        for i in range(len(word)):
-            cell = word[i]
-
-            if cell.letter is None:
-                    return 0
-            
-            if cell.state == False:
-                 counter = counter + cell.letter.value
-            
-            if cell.multiplier_type == 'letter' and cell.state == True:
-                counter = counter + (cell.letter.value * cell.multiplier * word_multiplier)
-
-            if cell.multiplier_type == 'word' and word_multiplier != 1 and cell.state == True:
-                counter = (counter + (cell.letter.value * word_multiplier))
-                word_multiplier = cell.multiplier
-                counter = counter * word_multiplier
-
-            if cell.multiplier_type == 'word' and word_multiplier == 1 and cell.state == True:
-                word_multiplier = cell.multiplier
-                counter = ((counter + cell.letter.value) * word_multiplier)
-            cell.used_cell(cell)
-        return counter       
+